Feat: Add UserRegisteredNotification listener dddd
This commit introduces a new event listener
`UserRegisteredNotification` in the `app/Listeners`
directory. The listener is responsible for handling
the `UserRegistered` event, which is triggered when
a new user registers in the system. Upon receiving
the event, the listener creates a notification
instance with details about the new user registration,
including the user's name. It then persists the
notification and creates user notifications for users
belonging to the membership affair office
(office_id = 2) to notify them about the new user
registration.