Fix: Make admin column nullable in admins table migration
This commit updates the `admins` table migration to
make the `admin` column nullable. Previously, it was
defined as unsignedBigInteger, which implies it
cannot be null. However, in certain cases, such as
when a user is not an admin but a regular member of
an office, the admin column should be nullable.

By making the admin column nullable, we allow for
more flexibility in defining admin status for users
in the context of different offices within the
application.