Add functionality to save the user list as a PDF
This commit introduces a new feature to the UserList.vue component, allowing users to save the entire list of users as a PDF file. A "Save as PDF" button is added to the interface, triggering the saveUsersAsPdf method when clicked. The method fetches all users' data asynchronously using the loadAllUsers function, prepares the data for the PDF table using the getUsersBodyContent helper method, and generates the PDF using the createPDFTemplate function from the pdfTemplate.js module. Finally, the PDF is saved with the filename "Users List.pdf"