Add route for fetching all users data
This commit extends the routes in the web.php file to include a new endpoint for fetching all users' data. The route /api/all-users is defined, mapped to the getAllUsers method within the UserController class. This endpoint is specifically created to support the new functionality added in the UserList.vue component, enabling users to save the entire list of users as a PDF file.